optimize and fixes some bug

+++ CookPopularControl/Tools/Interop/NativeMethods.cs
@@ -63,10 +63,10 @@ namespace CookPopularControl.Tools.Interop
         internal static extern uint GetWindowLong(IntPtr hWnd, int nIndex);
 
         [DllImport("user32.dll", SetLastError = true)]
-        internal static extern uint SetWindowLong(IntPtr hWnd, int nIndex, uint dwNewLong);
+        internal static extern uint SetWindowLong(IntPtr hWnd, int nIndex, int dwNewLong);
 
         [DllImport("user32.dll", SetLastError = true)]
-        internal static extern bool SetWindowPos(IntPtr hWnd, int hWndInsertAfter, int x, int y, int cx, int cy, SWP uFlags);
+        internal static extern bool SetWindowPos(IntPtr hWnd, IntPtr hWndInsertAfter, int x, int y, int cx, int cy, SWP uFlags);
 
         [DllImport("user32.dll", SetLastError = true)]
         [return: MarshalAs(UnmanagedType.Bool)]
